Calendar children filter plugged in
